Throw exception when more than one element is returned

+++ src/main/java/com/diffmin/DeletePatch.java
@@ -1,5 +1,6 @@
 package com.diffmin;
 
+import java.util.List;
 import spoon.reflect.CtModel;
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.path.CtPath;
@@ -11,6 +12,15 @@ public class DeletePatch {
 
     private CtModel modelToBePatched;
     private CtElement removedNode;
+
+    /**
+     * Custom exception when more than one element is returned.
+     */
+    class WrongNumberOfElementReturned extends Exception {
+        WrongNumberOfElementReturned(String message) {
+            super(message);
+        }
+    }
 
     /**
      * Set up node removal.
@@ -25,10 +35,18 @@ public class DeletePatch {
     /**
      * Removes `removedNode` from CtElement tree.
      */
-    public void process() {
+    public void process() throws WrongNumberOfElementReturned {
         CtPath removedNodePath = this.removedNode.getPath();
         // get(0) is used under that assumption that only one element will be returned
         // corresponding to a CtPath
-        removedNodePath.evaluateOn(this.modelToBePatched.getRootPackage()).get(0).delete();
+        List<CtElement> elementsToBeDeleted = removedNodePath.evaluateOn(
+            this.modelToBePatched.getRootPackage()
+        );
+        if (elementsToBeDeleted.size() == 1) {
+            removedNodePath.evaluateOn(this.modelToBePatched.getRootPackage()).get(0).delete();
+        }
+        else if (elementsToBeDeleted.size() > 1) {
+            throw new WrongNumberOfElementReturned("More than one element is found for deletion");
+        }
     }
 }
